Bug 989499 - mingw 64-bit fixup: use proper type in pointer cast to avoid losing precision.
UltraBlame original commit: 9c325a82110654cba3ee6e526b1fe5ab2920347b

+++ mozglue/misc/StackWalk.cpp
@@ -1073,7 +1073,7 @@ FramePointerStackWalk(MozWalkStackCallback aCallback, uint32_t aSkipFrames,
     
     if (next <= bp ||
         next > aStackEnd ||
-        (long(next) & 3)) {
+        (uintptr_t(next) & 3)) {
       break;
     }
 #if (defined(__ppc__) && defined(XP_MACOSX)) || defined(__powerpc64__)
